Another doh
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24445 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/fcp/ClientPutMessage.java
@@ -303,7 +303,8 @@ public class ClientPutMessage extends DataCarryingMessage {
 		if(persistenceType == ClientRequest.PERSIST_FOREVER)
 			container.activate(bucket, 5);
 		bucket.free();
-		bucket.removeFrom(container);
+		if(persistenceType == ClientRequest.PERSIST_FOREVER)
+			bucket.removeFrom(container);
 	}
 	
 }
